Move mint command into own file
